requirements.txt
Django==4.2.16
gunicorn==22.0.0
requests==2.31.0
python-dotenv==1.0.1
cloudinary==1.40.0
supabase==2.5.0
Pillow==10.3.0
pandas==2.2.2
geopy==2.4.1
folium==0.16.0
googlemaps==4.10.0
whitenoise==6.7.0